[add] : 편지 받기, 발송하기 기능

+++ app/src/main/java/com/coffeeinjection/message/data/remote/dto/MessageDtos.kt
@@ -9,7 +9,7 @@ import com.squareup.moshi.JsonClass
  */
 @JsonClass(generateAdapter = true)
 data class LetterSummaryDto(
-    @Json(name = "letterId") val letterId: Int,
+    @Json(name = "letterId") val letterId: Long,
     @Json(name = "content") val content: String,
     @Json(name = "senderName") val senderName: String,
     @Json(name = "status") val status: String,
